requirement.txt
python-whois
PyMySQL
coloredlogs
